millionx-phase2/stream-processors/requirements.txt
# Stream Processors - Python Dependencies

# Faust Streaming
faust-streaming==0.10.0

# Kafka Client (required by Faust)
kafka-python==2.0.2

# Data Validation
pydantic==2.5.0

# Redis for caching
redis==5.0.1

# HTTP requests (for weather API)
aiohttp==3.9.1

# Sentence Transformers for embeddings
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.36.0

# Weaviate vector database client
weaviate-client==3.26.0

# Monitoring and metrics
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
